refactor(#188): move code to another class
